Debug ProtocolAlreadyTerminated tests failure (#440)
This PR addresses issue #304, in which a possible unfavorable message ordering causes errors. Specifically, it was possible for a protocol to terminate locally before receiving a Ready message from yourself, causing the receipt of such message to trigger an error. To address this, the purpose of Ready messages was changed: instead of needing to receive a Ready from all players prior to starting a protocol instance (which is not required in the paper), we instead become "Ready" after receiving a Ready message from ourselves (so in practice, whoever is controlling the Participant can make them ready to run a protocol by sending a message).

This by itself doesn't fully fix the issue, so the other change is to not process any messages before receiving this initial Ready. Code was added to bank all messages received prior to Ready, and to play them back afterwards. To facilitate this, some functions were added to the Participant and MessageQueue API's.

To test that this change is effective, it was first implemented in KeyGen, which was tested for 20,000 consecutive runs without any issues. After that, the same changes were applied to Auxinfo and Presign. The end-to-end protocol test was run 200 times locally without issue.